Rename http package to http_requests
